Lots of changes
So basically i added a method to connect to the db in react then realised that would be insecure so made a separate part for the dbserver but will need to update db server and Jenkins to do a sparse checkout of just the DBSERVER folder. environment variables are on the db server already. other react things, added CORS and jwt.